requirements.txt
boto3==1.28.25
Flask==2.2.2
numpy==1.24.3
pandas==1.5.3
psycopg2==2.9.7
Werkzeug==2.2.2
regex==2023.6.3
pytz
rapidfuzz==3.1.2
google-generativeai
Flask-Cors==3.0.10
gunicorn==20.1.0